Change auto_peers default back to false and mark as deprecated.
We've seen some problems in production where time gets out of sync if
the number of peers is substantially more than the number of higher
stratum servers.  So auto_peers is unsuitable for environments with
large numbers of units.  Deprecated it in favour of the ntpmaster charm
which is designed for this purpose.